TRIX-446 moving <cloud>_controller to all.yml to make sure luna et al knows if we do have controllers in the cloud

+++ site/roles/cloud/luna/files/part.rd
@@ -1,14 +1,17 @@
 DISK=/dev/sda
+ROOTTHERE=$(parted $DISK print|grep -A100 Number|grep -i root|awk '{ print $1 }')
 
-for partition in $(parted $DISK print|grep -A100 Number|grep -v -e msftdata -e boot -e Number|awk '{ print $1 }'|grep -v '^$'); do
-    parted $DISK rm $partition
-done
+if [ ! "$ROOTTHERE" ]; then
+    for partition in $(parted $DISK print|grep -A100 Number|grep -v -e msftdata -e boot -e Number|awk '{ print $1 }'|grep -v '^$'); do
+        parted $DISK rm $partition
+    done
 
-parted $DISK -s 'mkpart root ext4 4g 100%'
+    parted $DISK -s 'mkpart root ext4 4g 100%'
+fi
 
-NEXT=$(parted $DISK print|grep -A100 Number|grep root|awk '{ print $1 }')
+ROOTPT=$(parted $DISK print|grep -A100 Number|grep -i root|awk '{ print $1 }')
 
-while [[ ! -b ${DISK}${NEXT} ]]; do sleep 1; done
+while [[ ! -b ${DISK}${ROOTPT} ]]; do sleep 1; done
  
-mkfs.ext4 ${DISK}${NEXT}
-mount ${DISK}${NEXT} /sysroot
+mkfs.ext4 ${DISK}${ROOTPT}
+mount ${DISK}${ROOTPT} /sysroot
